+++ app/Http/Services/UserService.php
@@ -5,9 +5,11 @@ namespace App\Http\Services;
 use App\Models\User;
 use App\Http\Resources\User\UserCollection;
 use App\Http\Resources\User\UserResource;
+use Exception;
 use Spatie\QueryBuilder\QueryBuilder;
-use  Illuminate\Http\Request;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
 
 final class UserService
 {
@@ -39,34 +41,70 @@ final class UserService
 
     public function find(int $id): UserResource
     {
+        $user = User::find($id);
+
+        if (!$user) throw new Exception('User not found');
+
         return new UserResource(User::find($id));
     }
 
     public function delete(int $id): void
     {
         $user = User::find($id);
+
+        if (!$user) throw new Exception('User not deleted');
+
         $user->delete();
     }
 
     public function destroy(int $id): void
     {
         $user = User::find($id);
+
+        if (!$user) throw new Exception('User not destroyed');
+
         $user->destroy();
     }
 
     public function restore(int $id): void
     {
-        User::withTrashed()->where('id', $id)->restore();
+        $user = User::withTrashed()->where('id', $id);
+
+        if (!$user) throw new Exception('User not restored');
+
+        $user->restore();
     }
 
     public function update(int $id, array $data): void
     {
+        $this->find($id);
+
+        $valid = Validator::make($data, [
+            'username' => 'required|max:255',
+            'firstname' => 'required|max:255',
+            'lastname' => 'required|max:255',
+            'password' => 'required|max:255',
+            'email' => 'required|max:255|unique:users',
+        ]);
+
+        if ($valid->fails()) throw new Exception('User not updated');
+
         User::where("id", $id)->update($data);
     }
 
     public function save(array $data): void
     {
-        if(isset($data['password'])) $data['password'] = Hash::make($data['password']);
+        $valid = Validator::make($data, [
+            'username' => 'required|max:255',
+            'firstname' => 'required|max:255',
+            'lastname' => 'required|max:255',
+            'password' => 'required|max:255',
+            'email' => 'required|max:255|unique:users',
+        ]);
+
+        if ($valid->fails()) throw new Exception('User not created');
+
+        if (isset($data['password'])) $data['password'] = Hash::make($data['password']);
 
         User::create($data);
     }
